feat: Create Mapper
