+++ app/src/main/java/com/b1nd/alimo/data/model/JoinModel.kt
@@ -1,4 +1,4 @@
-package com.b1nd.alimo.presentation.feature.onboarding.parent.join.third
+package com.b1nd.alimo.data.model
 
 data class JoinModel(
     val accessToken: String? = null,
