+++ app/src/main/java/com/b1nd/alimo/data/model/TokenModel.kt
@@ -1,15 +1,7 @@
 package com.b1nd.alimo.data.model
 
-import com.b1nd.alimo.data.local.entity.TokenEntity
-
 data class TokenModel(
     val token: String?,
     val refreshToken: String?
 )
 
-internal fun TokenEntity.toModel(): TokenModel {
-    return TokenModel(
-        token = token,
-        refreshToken = refreshToken
-    )
-}
