+++ app/src/main/java/com/b1nd/alimo/data/remote/mapper/TokenMapper.kt
@@ -1,0 +1,11 @@
+package com.b1nd.alimo.data.remote.mapper
+
+import com.b1nd.alimo.data.local.entity.TokenEntity
+import com.b1nd.alimo.data.model.TokenModel
+
+internal fun TokenEntity.toModel(): TokenModel {
+    return TokenModel(
+        token = token,
+        refreshToken = refreshToken
+    )
+}
